# docs/marinelink_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MarineLink B2B Seafood API",
    "description": "API Collection for MarineLink seafood wholesaling application.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Spring Boot Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AdminProductRequest": {
        "type": "object",
        "required": ["name", "categoryId", "basePrice", "unit", "minOrderQuantity", "stockQuantity", "status"],
        "properties": {
          "name": { "type": "string", "example": "Muc kho loai 1" },
          "categoryId": { "type": "string", "format": "uuid" },
          "basePrice": { "type": "number", "format": "double", "example": 450000 },
          "unit": { "type": "string", "example": "kg" },
          "minOrderQuantity": { "type": "integer", "minimum": 1, "example": 2 },
          "stockQuantity": { "type": "integer", "minimum": 0, "example": 120 },
          "status": { "type": "string", "enum": ["ACTIVE", "OUT_OF_STOCK", "DISABLED"], "example": "ACTIVE" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "List seafood catalog",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Search keyword" },
          { "name": "categoryId", "in": "query", "schema": { "type": "string", "format": "uuid" }, "description": "Category public UUID. Parent category IDs include products from direct child categories." },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["ACTIVE", "OUT_OF_STOCK", "DISABLED"] } },
          { "name": "featured", "in": "query", "schema": { "type": "boolean" } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["newest", "price_asc", "price_desc", "name_asc", "name_desc"] }, "description": "Whitelisted product sort key" }
        ],
        "responses": {
          "200": {
            "description": "Seafood list"
          }
        }
      }
    },
...
    "/api/products/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get product detail including wholesale tiers",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "550e8400-e29b-41d4-a716-446655440111" }
        ],
        "responses": {
          "200": {
            "description": "Product details"
          }
        }
      }
    },
...
    "/api/admin/products/{id}": {
      "get": {
        "tags": ["Admin Portal"],
        "summary": "Get product detail for admin management",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Admin product detail"
          }
        }
      },
      "put": {
        "tags": ["Admin Portal"],
        "summary": "Update product",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdminProductRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated"
          }
        }
      },
      "delete": {
        "tags": ["Admin Portal"],
        "summary": "Soft delete or disable product",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Product disabled"
          }
        }
      }
    },
...
  }
}